# [SPARK-35879][Core][Shuffle] Fix performance regression caused by collectFetchRequests