@mridulm 2021-06-25
Thanks for the changes @yaooqinn and thanks for the ping @dongjoon-hyun :-)
The changes look good - but I want to make sure we still have similar performance benefits as we saw in the PR description.
Will it be possible to rerun the benchmark to ensure no regression @yaooqinn ? Thx !