@mridulm 2021-06-25
@dongjoon-hyun I am +1 on the change if the performance is still good based on @yaooqinn's benchmark (will be going on vacation soon, so dont want reviews to be blocked on my accord !). Thx